+++ ply-plutarch/src/Ply/Plutarch.hs
@@ -1,4 +1,4 @@
-module Ply.Plutarch (PlyArgOf, writeTypedScript, mkTypedScript) where
+module Ply.Plutarch (PlyArgOf, writeTypedScript, mkEnvelope) where
 
 import Ply.Plutarch.Class (PlyArgOf)
-import Ply.Plutarch.TypedWriter (mkTypedScript, writeTypedScript)
+import Ply.Plutarch.TypedWriter (mkEnvelope, writeTypedScript)